# konnect: handle deletions first