@pmalek 2024-10-29
This would require the change in behavior described in https://github.com/Kong/gateway-operator/pull/821: so that the operator does not set the ownership relations between object and objects living inside the CP are not scheduled for deletion when CP gets deleted.

Closing until that changes.